+++ src/neuroagent/tools/electrophys_tool.py
@@ -7,8 +7,8 @@ from typing import Any, Literal, Optional, Type
 
 from bluepyefe.extract import extract_efeatures
 from efel.units import get_unit
-from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.tools import ToolException
+from pydantic import BaseModel, Field
 
 from neuroagent.tools.base_tool import BaseToolOutput, BasicTool
 from neuroagent.utils import get_kg_data