+++ src/neuroagent/tools/morphology_features_tool.py
@@ -5,9 +5,9 @@ from typing import Any, Type
 
 import neurom
 import numpy as np
-from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.tools import ToolException
 from neurom.io.utils import load_morphology
+from pydantic import BaseModel, Field
 
 from neuroagent.tools.base_tool import BaseToolOutput, BasicTool
 from neuroagent.utils import get_kg_data