+++ CHANGELOG.md
@@ -7,5 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Github action to create the docs.
 
-
+### Changed
+- Migration to pydantic V2.